+++ sdk/core/azure-core-tracing-opentelemetry/CHANGELOG.md
@@ -15,6 +15,7 @@
     - `http.url` -> `url.full`
 - Removed `http.user_agent` optional attribute since the same information is reported in the instrumentation scope via library name and version.  
 - Removed `OpenTelemetrySchemaVersion` and it's setter method on `OpenTelemetryTracingOptions` since we're not allowing to change the schema version for now.
+- Removed exception event reporting - exceptions are reported as logs already.
 - 
 ### Bugs Fixed